Proof-of-stake updated % reward for continuous Minter as per 80eb9cd
PoS rewards are given to those securing the network against attackers and are determined by a linear function with two components. The first dynamic portion results in 3% of the coins used to mint annually. The second static portion results in an additional 0.25% inflation of the coin supply distributed evenly amongst the blocks expected in a year, which approximately equals 1.25 ppc but scales with the size of the supply.

Proof-of-Stake reward formula is:
Proof-of-Stake coin mint rate is ~3-5% interest per year